// packages/cashconnect/src/core/types/controllerTypes.ts
export type WalletNetwork =
  | 'mainnet'
  | 'chipnet'
  | 'testnet3'
  | 'testnet4'
  | 'regtest'

export type WalletProtocol = 'bip70' | 'jpp' | 'jppv2' | 'wc2'
export interface DesktopWallet {
  id: string
  name: string
  links?: {
    native: string | (() => string) | (() => Promise<string>)
    universal?: string | (() => string) | (() => Promise<string>)
  }
  disableLinkFormatting?: boolean
  image?: string
  networks: WalletNetwork[]
  protocols: WalletProtocol[]
}

export type Platform =
  | 'desktop'
  | 'web'
  | 'ios'
  | 'android'
  | 'chrome-extension'

export interface Wallet {
  id: string
  name: string
  links?: {
    native: string | (() => string) | (() => Promise<string>)
    universal?: string | (() => string) | (() => Promise<string>)
  }
  disableLinkFormatting?: boolean
  image?: string
  networks: WalletNetwork[]
  protocols: WalletProtocol[]
  platforms: Platform[]
  disableOnMobile?: boolean
}
// ...
export interface WalletData {
  id: string
  name: string
  homepage?: string
  image_id?: string
  image?: string
  app?: {
    browser?: string
    ios?: string
    android?: string
  }
  mobile?: {
    native?: string | (() => string) | (() => Promise<string>)
    universal?: string | (() => string) | (() => Promise<string>)
  }
  desktop?: {
    native?: string | (() => string) | (() => Promise<string>)
    universal?: string | (() => string) | (() => Promise<string>)
  }
  disableLinkFormatting?: boolean
  networks: WalletNetwork[]
  protocols: WalletProtocol[]
  disableOnMobile?: boolean
}